Auto generatedId

+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/domain/ReferenceRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReferenceRepository extends JpaRepository<Reference, Long> {
+public interface ReferenceRepository extends JpaRepository<Reference, Integer> {
 }
